add a test of arg parsing
